requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0          # async PostgreSQL driver
psycopg2-binary==2.9.9   # sync driver (for Alembic migrations)

# Validation & settings
pydantic==2.7.1
pydantic-settings==2.3.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File uploads
python-multipart==0.0.9

# QR code generation
qrcode[pil]==7.4.2
pillow==10.3.0

# AWS S3
boto3==1.34.110

# HTTP client (for external APIs)
httpx==0.27.0

# PDF manifest generation
reportlab==4.2.0

# Utilities
python-slugify==8.0.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0            # async test client
factory-boy==3.3.0
aiosqlite==0.20.0
greenlet
